Fixed following issues
1. Searching by name does not work correctly

Root cause:- Due to case mismatch, it was not finding record correctly.
added uppercase() on either side so that it will always compare uppercase
character on either side.

2.When no player is found, a 500 error occurs.
3. When no player is entered and submit is pressed, a 500 error occurs.

Root cause:-

It was giving 500(internal server) because we were not handling null/empty condition
of class. It has been fixed now.